refactor(Event): Event가 뒤로 오도록 클래스명 수정
