+++ src/main/java/christmas/domain/event/GiveawayEvent.java
@@ -5,12 +5,12 @@ import christmas.domain.event.eventdate.EventDate;
 import christmas.domain.event.giveaway.Giveaway;
 import christmas.domain.menu.OrderMenu;
 
-public class EventGiveaway implements Event {
+public class GiveawayEvent implements Event {
     private final String eventName;
     private final Giveaway giveaway;
     private final EventDate eventDate;
 
-    public EventGiveaway(String eventName, Giveaway giveaway, EventDate eventDate) {
+    public GiveawayEvent(String eventName, Giveaway giveaway, EventDate eventDate) {
         this.eventName = eventName;
         this.giveaway = giveaway;
         this.eventDate = eventDate;
