+++ src/test/java/christmas/domain/event/DiscountEventTest.java
@@ -1,8 +1,6 @@
-package christmas.domain.event.discount;
+package christmas.domain.event;
 
 import christmas.domain.Money;
-import christmas.domain.event.EventBenefitDetail;
-import christmas.domain.event.EventDiscount;
 import christmas.domain.menu.Menu;
 import christmas.domain.menu.MenuAmount;
 import christmas.domain.menu.OrderMenu;
@@ -11,7 +9,7 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-class EventDiscountTest {
+class DiscountEventTest {
 
     @Test
     @DisplayName("할인, 일정, 이벤트 이름을 바탕으로 할인 정보를 반환한다.")
@@ -22,7 +20,7 @@ class EventDiscountTest {
                         new MenuAmount(Menu.ICE_CREAM, 1)
                 )
         );
-        EventDiscount alwaysDiscount = new EventDiscount("무조건 천원 할인", (orders -> Money.of(1_000L)), (date) -> true);
+        DiscountEvent alwaysDiscount = new DiscountEvent("무조건 천원 할인", (orders -> Money.of(1_000L)), (date) -> true);
 
         // when
         EventBenefitDetail eventBenefitDetail = alwaysDiscount.getBenefitDetail(orderMenu);
